test/sim,msvc: fix signed-overflow UB in the app-hash fold; surface cl errors
Both found by CI on the v1.49.2 push -- the exact case AGENTS.md warns
about, where a tier the local run cannot reach catches what local green
misses.

SIGNED OVERFLOW (the new sanitizers-dst job, which this release added,
doing its job).  Every sim test folds an order-sensitive hash with
`h = h * 1000003L + ...` on a signed `long`.  Wrapping is the INTENT,
but signed overflow is undefined behavior, so UBSan flagged six tests
(sched, partition, buggify, iofault, torn, soak):

    runtime error: signed integer overflow:
    2000024000100000139 * 1000003 cannot be represented in type 'long int'

PRE-EXISTING, not from this release's work: the idiom is present at
v1.48.0 and v1.49.1, and none of the review commits introduced it.  It
only became visible because ASan/UBSan previously ran the NATIVE suite
while the DST harness compiled its tests WITHOUT sanitizer flags, so
nothing ever put UBSan and the sim tests in the same build.

Fixed at all 37 sites in 30 files by doing the multiply in UNSIGNED
arithmetic (where wrapping is defined) and converting back.  Deliberate
over -fwrapv or a narrower type: it is local to the tests, needs no
build-flag coordination, and -- critically -- is VALUE-PRESERVING, which
a replay-hash change would not be.  Verified rather than assumed: a
standalone probe shows the unsigned path yields identical wrapped values,
and test_sim_sched reports state=eb0ad45b150118e5 /
hash=4594351820588205126 byte-identically before and after.  That matters
because the pinned corpus and every "replays identically" assertion are
keyed on these hashes; had they shifted, the regression ledger would have
been silently invalidated.

MSVC DIAGNOSABILITY.  The same run reported `m14/test_cfg BUILD FAILED`
with NO error text, because all three build sites in dist/build_msvc.bat
send cl output to >nul.  A build break on a platform with no local host
is exactly when that text is needed, and its absence cost a round trip
here.  All three sites now tee to a log and `type` it on failure (the
advisory site included, kept non-gating).

The two MSVC failures themselves (m14/test_cfg build, m11/test_slab run)
are NOT fixed by this commit -- they need the error text the next CI run
will produce.  Both are regressions from this release's work: each was
passing and gated at v1.49.1, so they are ours, not pre-existing.

Verified: six formerly-failing tests UBSan-clean (rc=0, 0 runtime
errors); make check-dst 67/67 rc=0; dst-corpus 16/16 pinned seeds still
pass.

+++ test/sim/test_sim_aiov.c
@@ -59,7 +59,8 @@ fold(long v)
 	do {
 		h = atomic_load_explicit(&g_hash, memory_order_relaxed);
 	} while (!atomic_compare_exchange_weak_explicit(&g_hash, &h,
-	    h * 1000003L + (v + 11), memory_order_relaxed,
+	    (long)((unsigned long)h * 1000003UL +
+	    (unsigned long)(v + 11)), memory_order_relaxed,
 	    memory_order_relaxed));
 }
