+++ test/sim/test_sim_buggify3.c
@@ -74,7 +74,8 @@ static void
 a_fold(int v)
 {
 	long h = atomic_load_explicit(&g_a_hash, memory_order_relaxed);
-	h = h * 1000003L + (v + 1);
+	h = (long)((unsigned long)h * 1000003UL +
+		    (unsigned long)(v + 1));
 	atomic_store_explicit(&g_a_hash, h, memory_order_relaxed);
 }
 
@@ -194,7 +195,8 @@ static void
 b_fold(long v)
 {
 	long h = atomic_load_explicit(&g_b_hash, memory_order_relaxed);
-	h = h * 1000003L + (v + 1);
+	h = (long)((unsigned long)h * 1000003UL +
+		    (unsigned long)(v + 1));
 	atomic_store_explicit(&g_b_hash, h, memory_order_relaxed);
 }
