+++ test/sim/test_sim_dispatch.c
@@ -46,7 +46,8 @@ effect_body(void *arg)
 	long h;
 	atomic_fetch_add_explicit(&g_ran, 1, memory_order_relaxed);
 	h = atomic_load_explicit(&g_hash, memory_order_relaxed);
-	h = h * 1000003L + (id + 1);
+	h = (long)((unsigned long)h * 1000003UL +
+		    (unsigned long)(id + 1));
 	atomic_store_explicit(&g_hash, h, memory_order_relaxed);
 	return (int)id;
 }
