+++ test/sim/test_sim_rcu.c
@@ -88,7 +88,8 @@ static void
 rcu_fold(long v)
 {
 	long h = atomic_load_explicit(&g_hash, memory_order_relaxed);
-	h = h * 1000003L + (v + 1);
+	h = (long)((unsigned long)h * 1000003UL +
+		    (unsigned long)(v + 1));
 	atomic_store_explicit(&g_hash, h, memory_order_relaxed);
 }
