+++ test/sim/test_sim_swarm.c
@@ -254,7 +254,8 @@ torn_verifier(void *arg)
 		{
 			long h = atomic_load_explicit(&g_app_hash,
 			    memory_order_relaxed);
-			h = h * 1000003L + (id + 100);
+			h = (long)((unsigned long)h * 1000003UL +
+		    (unsigned long)(id + 100));
 			atomic_store_explicit(&g_app_hash, h,
 			    memory_order_relaxed);
 		}
@@ -331,7 +332,8 @@ ping(void *arg)
 		atomic_fetch_add_explicit(&g_pair_replies[pa->id], 1,
 		    memory_order_relaxed);
 		h = atomic_load_explicit(&g_app_hash, memory_order_relaxed);
-		h = h * 1000003L + (pa->id + 1);
+		h = (long)((unsigned long)h * 1000003UL +
+		    (unsigned long)(pa->id + 1));
 		atomic_store_explicit(&g_app_hash, h, memory_order_relaxed);
 		hops--;
 	}
